부모 대시보드, 심리 설문 관련 페이지 추가: 부모 대시보드, 심리 설문 관련 페이지를 구현합니다.

+++ src/main/java/com/scit/iLog/domain/mentalsurvey/QuestionResponse.java
@@ -9,5 +9,5 @@ import lombok.*;
 public class QuestionResponse {
     private String questionItem;
     private String answer;
-    private double score;
+    private int score;
 }
